функция getCommentObj  из util перенесена в data
